requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
python-dotenv>=1.0.0
redis>=4.5.0
sqlalchemy>=2.0.0
alembic>=1.10.0

# Use psycopg2-binary only when needed
# psycopg2-binary>=2.9.5

# LLM and vector DB
openai>=1.0.0
anthropic>=0.5.0

# ChromaDB and LangChain - commented out for initial setup
# chromadb>=0.4.0
# langchain>=0.0.200

# Gateway integrations - using standard libraries instead
# python-whatsapp>=0.1.0  # replaced with custom WhatsApp Business API implementation
# python-gmail>=0.1.0  # using built-in email libraries

# HTTP client for external APIs
aiohttp>=3.8.0

# Email libraries
email-validator>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0

# Dev tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Monitoring
prometheus-client>=0.16.0
opentelemetry-api>=1.15.0
opentelemetry-sdk>=1.15.0

# Logging
python-json-logger>=2.0.7
